Move the load data code into its own file in lib
